+++ src/main/java/com/ecam/picto/pictopro/controller/ProfessionnelController.java
@@ -43,7 +43,7 @@ public class ProfessionnelController {
 	}
 
 	@GetMapping(path = "/{id}")
-	public String afficherDetailsPro(@PathVariable("id") int id, Model model) {
+	public String afficherDetailsPro(@PathVariable("id") Long id, Model model) {
 		Optional<Professionnel> professionnel = this.professionnelRepository.findById(id);
 		if (professionnel.isEmpty()) {
 			throw new ResponseStatusException(HttpStatus.NOT_FOUND, "entity not found");
